worked out how to use LINQ
to query/filter/sort data. LINQ can also be used when data is being retrieved probably - it allows us to do SQL style queries on the working copies of the data, instead of querying the tables service. This might mean that the best approach when filtering results finely is to grab a coarser subset of the data from the azure table and finish digesting it in the controller layer or w/e.